+++ src/playground_base/playground_base/base/topo/static/FBScalarParamAddrs.hpp
@@ -2,6 +2,6 @@
 
 #include <functional>
 
-typedef std::function<float* (
+typedef std::function<double* (
   int moduleSlot, int paramSlot, void* state)>
 FBScalarParamAddrSelector;
